#42 configure subspec
