+++ NodeKit.podspec
@@ -15,8 +15,8 @@ Pod::Spec.new do |s|
   s.default_subspec = 'Core'
 
   s.subspec 'Core' do |sp|
-    sp.source_files  = 'NodeKit/Utils/**/*.swift', 'NodeKit/Chains/**/*.swift', 'NodeKit/Layers/**/*.swift', 'NodeKit/Core/**/*.swift', 'NodeKit/Encodings/*.swift'
-    sp.dependency 'Alamofire', '5.0.0-beta.4'
+    sp.source_files  = 'NodeKit/Utils/**/*.swift', 'NodeKit/Chains/**/*.swift', 'NodeKit/Layers/**/*.swift', 'NodeKit/Core/**/*.swift', 'NodeKit/Encodings/*.swift', 'NodeKit/ThirdParty/**/*.swift'
+    sp.exclude_files = 'NodeKit/Encodings/UrlBsonRequestEncodingNode.swift', 'NodeKit/Chains/UrlBsonServiceChainBuilder.swift', 'NodeKit/Chains/UrlBsonChainsBuilder.swift', 'NodeKit/Layers/Bson/*.swift', 'NodeKit/Layers/TransportLayer/Models/TransportUrlBsonRequest.swift', 'NodeKit/Layers/ResponseProcessingLayer/Bson/*.swift', 'NodeKit/Layers/BsonLayerTypes.swift', 'NodeKit/Core/Convertion/Extensions/DTOConvertible+Document.swift', 'NodeKit/Core/Convertion/RawBsonMappable.swift'
     sp.dependency 'CoreEvents', '~> 2.0.1'
     sp.dependency 'BSON', '7.0.4'
   end
@@ -31,4 +31,9 @@ Pod::Spec.new do |s|
     sp.source_files = 'NodeKit/CacheNode/**/*.swift'
   end
 
+  s.subspec 'BSON' do |sp|
+  	sp.dependency 'NodeKit/Core'
+  	sp.dependency 'BSON', '7.0.4'
+  	sp.source_files = 'NodeKit/Encodings/UrlBsonRequestEncodingNode.swift', 'NodeKit/Chains/UrlBsonServiceChainBuilder.swift', 'NodeKit/Chains/UrlBsonChainsBuilder.swift', 'NodeKit/Layers/Bson/*.swift', 'NodeKit/Layers/TransportLayer/Models/TransportUrlBsonRequest.swift', 'NodeKit/Layers/ResponseProcessingLayer/Bson/*.swift', 'NodeKit/Layers/BsonLayerTypes.swift', 'NodeKit/Core/Convertion/Extensions/DTOConvertible+Document.swift', 'NodeKit/Core/Convertion/RawBsonMappable.swift'
+  end
 end
